+++ app/mediaPlayer/Util/playerMsgQ.h
@@ -12,7 +12,7 @@ class playerMsgQ
     playerMsgQ();
     ~playerMsgQ();
 
-    void InitComponent(int totalMsgSizeLimit);
+    void InitComponent(size_t totalMsgSizeLimit);
     bool AddMsg(std::shared_ptr<PlayerMsg_Base> msg);
     void GetMsg(std::shared_ptr<PlayerMsg_Base>& msg);
     int GetMsgNum();
@@ -20,7 +20,7 @@ class playerMsgQ
  private:
     DefaultMutex m_mutex;
     pthread_cond_t m_cond;
-    int m_totalMsgSizeLimit;
+    size_t m_totalMsgSizeLimit;
     int m_totalMsgSize;
     std::queue<std::shared_ptr<PlayerMsg_Base> > m_msgQ;
 };
